Fazendo um clean code básico
